Split UiUtils file
