--- src/ThreeEditor/js/object/Object.Filter.ts
@@ -1,7 +1,0 @@
-import { ObjectAbstract } from './Object.Abstract';
-
-export class ObjectFilter extends ObjectAbstract {
-	update(): void {
-		throw new Error('Method not implemented.');
-	}
-}
