+++ src/ThreeEditor/js/sidebar/object/Object.Dimensions.ts
@@ -6,25 +6,25 @@ import {
 	isBoxMesh,
 	isCylinderMesh,
 	isSphereMesh
-} from '../../util/BasicMeshes';
-import { DetectGeometry, isDetectGeometry } from '../../util/Detect/DetectGeometry';
-import { DETECT_OPTIONS } from '../../util/Detect/DetectTypes';
+} from '../../../util/BasicMeshes';
+import { DetectGeometry, isDetectGeometry } from '../../../util/Detect/DetectGeometry';
+import { DETECT_OPTIONS } from '../../../util/Detect/DetectTypes';
 import {
 	createRowParamNumber,
 	createRowSelect,
 	createRowText,
 	hideUIElement,
 	showUIElement
-} from '../../util/UiUtils';
-import { isWorldZone, WorldZone } from '../../util/WorldZone';
+} from '../../../util/Ui/Uis';
+import { isWorldZone, WorldZone } from '../../../util/WorldZone';
 import {
 	SetDetectGeometryCommand,
 	SetDetectTypeCommand,
 	SetGeometryCommand,
 	SetValueCommand
-} from '../commands/Commands';
-import { Editor } from '../Editor';
-import { UINumber, UIRow, UISelect, UIText } from '../libs/ui';
+} from '../../commands/Commands';
+import { Editor } from '../../Editor';
+import { UINumber, UIRow, UISelect, UIText } from '../../libs/ui';
 import { ObjectAbstract } from './Object.Abstract';
 
 export class ObjectDimensions extends ObjectAbstract {
