+++ src/ThreeEditor/js/sidebar/object/Object.Info.ts
@@ -1,7 +1,7 @@
-import { createRowParamInput, createRowText } from '../../util/UiUtils';
-import { SetValueCommand } from '../commands/Commands';
-import { Editor } from '../Editor';
-import { UIInput, UIRow, UIText } from '../libs/ui';
+import { createRowParamInput, createRowText } from '../../../util/Ui/Uis';
+import { SetValueCommand } from '../../commands/Commands';
+import { Editor } from '../../Editor';
+import { UIInput, UIRow, UIText } from '../../libs/ui';
 import { ObjectAbstract } from './Object.Abstract';
 
 export class ObjectInfo extends ObjectAbstract {
